+++ esp-hal-embassy/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Re-added the `multiple-integrated` timer queue flavour (#3166)
+
 ### Changed
 
 ### Fixed