+++ esp-hal-embassy/CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed a panic on very long wakeup times (#3433)
+
 ### Removed
 
 ## [0.7.0] - 2025-02-24